+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 the [PEP 440 version scheme](https://peps.python.org/pep-0440/#version-scheme).
 
+## [0.1.0-alpha2]
+### Changed
+- Use tuple instead of ListView for pytrified sequences.
+- pytrified Sequences are JSON serializable.
+
 ## [0.1.0b1]
 ### Added
 - pytrify()